+++ example/src/auto/auto_mode.ts
@@ -195,6 +195,8 @@ const CHAT_OPTIONS_KEYS = [
   'dohVendor',
   'enableUserInfo',
   'enableAutoSyncContacts',
+  'enableChatroomConversation',
+  'autoLoadConversations',
 ];
 
 function deriveInitParams(